requirements.txt
python-telegram-bot
transformers
torch
speechrecognition
gTTS
python-dotenv
scikit-learn